Changes for an initial npm package release
