+++ r2pipe/nodejs/package.json
@@ -1,0 +1,44 @@
+{
+  "name": "r2pipe",
+  "version": "0.1.0",
+  "description": "pipe bindings for radare2",
+  "main": "index.js",
+  "scripts": {
+    "test": "node test.js"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/radare/radare2-bindings.git"
+  },
+   "files": [
+    "README.md",
+    "index.js",
+    "test.js"
+  ],
+  "keywords": [
+    "radare",
+    "radare2",
+    "r2",
+    "reversing",
+    "disassembler",
+    "hexadecimal",
+    "editor",
+    "exploit",
+    "exploiting"
+  ],
+  "author": {
+    "name": "pancake",
+    "email": "pancake@nopcode.org"
+  },
+  "contributors": [
+    {
+      "name": "Jaime Peñalba",
+      "email": "jpenalbae@gmail.com"
+    }
+  ],
+  "license": "LGPLv3",
+  "bugs": {
+    "url": "https://github.com/radare/radare2-bindings/issues"
+  },
+  "homepage": "http://www.radare.org"
+}
